Refactored Note Sync, wrote tests
