+++ .vscode/launch.json
@@ -10,7 +10,8 @@
 			"type": "extensionHost",
 			"request": "launch",
 			"args": [
-				"--extensionDevelopmentPath=${workspaceFolder}"
+				"--extensionDevelopmentPath=${workspaceFolder}", 
+				"${workspaceFolder}/test/ws_manual/"
 			],
 			"outFiles": [
 				"${workspaceFolder}/dist/**/*.js"
@@ -23,7 +24,8 @@
 			"request": "launch",
 			"args": [
 				"--extensionDevelopmentPath=${workspaceFolder}",
-				"--extensionTestsPath=${workspaceFolder}/out/test/suite/index"
+				"--extensionTestsPath=${workspaceFolder}/out/test/suite/index", 
+				"${workspaceFolder}/test/ws_unittests/"
 			],
 			"outFiles": [
 				"${workspaceFolder}/out/**/*.js",
