+++ package.json
@@ -303,13 +303,13 @@
   "scripts": {
     "rimraf": "./node_modules/rimraf/bin.js", 
     "vscode:prepublish": "npm run package",
-    "compile": "npm run clean && webpack",
+    "compile": "webpack",
     "watch": "webpack --watch",
     "package": "webpack --mode production --devtool hidden-source-map",
-    "clean": "rimraf out/", 
+    "clean": "rimraf out/ && rimraf test/ws_unittests/journal", 
     "compile-tests": "tsc -p . --outDir out",
     "rebuild-tests": "tsc --build --clean tsconfig.json",
-    "watch-tests": "tsc -p . -w --outDir out",
+    "watch-tests": "npm run clean && tsc -p . -w --outDir out",
     "pretest": "npm run clean && npm run compile-tests && npm run compile && npm run lint",
     "lint": "eslint src --ext ts",
     "test": "node ./out/test/runTest.js"
