+++ src/actions/inject.ts
@@ -18,17 +18,12 @@
 
 'use strict';
 
-import * as Path from 'path';
+
 import * as vscode from 'vscode';
 import * as J from '../.';
 
 
-interface InlineString {
-    position: vscode.Position;
-    value: string;
-    document: vscode.TextDocument;
 
-}
 
 export class Inject {
 
@@ -58,21 +53,21 @@ export class Inject {
                     if (input.flags.match("memo")) {
                         this.ctrl.config.getMemoInlineTemplate()
                             .then(tplInfo => this.buildInlineString(doc, tplInfo, ["${input}", input.text]))
-                            .then((val: InlineString) => this.injectInlineString(val))
+                            .then((val: J.Model.InlineString) => this.injectInlineString(val))
                             .then(doc => resolve(doc))
                             .catch((err) => reject(err));
 
                     } else if (input.flags.match("task")) {
                         this.ctrl.config.getTaskInlineTemplate()
                             .then(tplInfo => this.buildInlineString(doc, tplInfo, ["${input}", input.text]))
-                            .then((val: InlineString) => this.injectInlineString(val))
+                            .then((val: J.Model.InlineString) => this.injectInlineString(val))
                             .then(doc => resolve(doc))
                             .catch((err) => reject(err));
 
                     } else if (input.flags.match("todo")) {
                         this.ctrl.config.getTaskInlineTemplate()
                             .then(tplInfo => this.buildInlineString(doc, tplInfo, ["${input}", input.text]))
-                            .then((val: InlineString) => this.injectInlineString(val))
+                            .then((val: J.Model.InlineString) => this.injectInlineString(val))
                             .then(doc => resolve(doc))
                             .catch((err) => reject(err));
                     } else {
@@ -106,7 +101,7 @@ export class Inject {
      * 
      * Updates: Fix for  #55, always make sure there is a linebreak between the header and the injected text to stay markdown compliant
      */
-    private async buildInlineString(doc: vscode.TextDocument, tpl: J.Extension.InlineTemplate, ...values: string[][]): Promise<InlineString> {
+    public async buildInlineString(doc: vscode.TextDocument, tpl: J.Model.InlineTemplate, ...values: string[][]): Promise<J.Model.InlineString> {
 
         return new Promise((resolve, reject) => {
 
@@ -169,16 +164,18 @@ export class Inject {
      * @param other additional InlineStrings
      * 
      */
-    public async injectInlineString(content: InlineString, ...other: InlineString[]): Promise<vscode.TextDocument> {
+    public async injectInlineString(content: J.Model.InlineString, ...other: J.Model.InlineString[]): Promise<vscode.TextDocument> {
 
 
         return new Promise((resolve, reject) => {
+            this.ctrl.logger.trace("Entering injectInlineString() in inject.ts with string: ", content.value.trim());
+
             if (J.Util.isNullOrUndefined(content)) {
                 this.ctrl.logger.error("Content is null");
-                return reject("Invalid call, no reference to document due to null content.");
+                reject("Invalid call, no reference to document due to null content.");
+                return;
             }
 
-            this.ctrl.logger.trace("Entering injectInlineString() in inject.ts with string: ", content.value.trim());
 
             try {
 
@@ -203,8 +200,11 @@ export class Inject {
                             if (applied === true) {
                                 resolve(content.document);
                             } else {
+                                this.ctrl.logger.error("Failed inject inline string '", content.value, "'");
                                 reject("Failed to applied edit");
                             }
+                        }, rejected => {
+                            this.ctrl.logger.error("Failed inject inline string, reason: ", rejected);
                         });
                 }
 
@@ -216,7 +216,7 @@ export class Inject {
         });
     }
 
-    private formatContent(content: InlineString) {
+    private formatContent(content: J.Model.InlineString) {
         if (J.Util.isNullOrUndefined(content.position)) {
             content.position = new vscode.Position(1, 0);
         }
@@ -278,7 +278,7 @@ export class Inject {
 
             // Fixme: add the tags inject them after header
             this.ctrl.config.getNotesTemplate(input.scope)
-                .then((ft: J.Extension.HeaderTemplate) => {
+                .then((ft: J.Model.HeaderTemplate) => {
                     ft.value = ft.value!.replace('${input}', input.text);
                     ft.value = ft.value!.replace('${tags}', input.tags.join(" ") + '\n');
 
@@ -291,108 +291,8 @@ export class Inject {
 
 
 
-    /**
-     * Injects a reference to a file associated with the given document. The reference location can be configured in the template (after-flag)
-     * @param doc the document which we will inject into
-     * @param file the referenced path 
-     */
-    private async buildReference(doc: vscode.TextDocument, file: vscode.Uri): Promise<InlineString> {
-        return new Promise<InlineString>((resolve, reject) => {
-            try {
-                this.ctrl.logger.trace("Entering injectReference() in ext/inject.ts for document: ", doc.fileName, " and file ", file);
-
-                this.ctrl.config.getFileLinkInlineTemplate()
-                    .then(tpl => {
-                        // fix for #70 
-                        const pathToLinkedFile: Path.ParsedPath = Path.parse(file.fsPath);
-                        const pathToEntry: Path.ParsedPath = Path.parse(doc.uri.fsPath);
-                        const relativePath = Path.relative(pathToEntry.dir, pathToLinkedFile.dir);
-                        const link = Path.join(relativePath, pathToLinkedFile.name + pathToLinkedFile.ext);
-
-                        let title = pathToLinkedFile.name.replace(/_/g, " ");
-                        if (pathToLinkedFile.ext.substr(1, pathToLinkedFile.ext.length) !== this.ctrl.config.getFileExtension()) {
-                            title = "(" + pathToLinkedFile.ext + ") " + title;
-                        };
 
 
-                        return this.buildInlineString(
-                            doc,
-                            tpl,
-                            ["${title}", title],
-                            // TODO: reference might refer to other locations 
-                            ["${link}", link]
-                        );
-                    }
-                    )
-                    .then(inlineString => resolve(inlineString))
-                    .catch(error => {
-                        this.ctrl.logger.error("Failed to inject reference. Reason: ", error);
-                        reject(error);
-                    });
-
-
-            } catch (error) {
-                reject(error);
-            }
-        });
-    }
-
-    /**
-     * Checks for the given text document if it contains references to notes (and if there are notes in the associated folders)
-     * It compares the two lists and creates (or deletes) any missing links
-     * 
-     * @param doc 
-     */
-    public async injectAttachementLinks(doc: vscode.TextDocument, date: Date): Promise<vscode.TextDocument> {
-        return new Promise((resolve, reject) => {
-            this.ctrl.logger.trace("Entering injectAttachementLinks() in inject.ts for date: ", date);
-
-
-            this.ctrl.ui.saveDocument(doc)
-                .then(() =>
-
-                    // FIXME: We have to change the logic here: first generate the link according to template, then check if the generated text is already in the document
-
-                    // we invoke the scan of the notes directory in parallel
-                    Promise.all([
-                        this.ctrl.reader.getReferencedFiles(doc),
-                        this.ctrl.reader.getFilesInNotesFolderAllScopes(doc, date)
-                    ])
-                )
-                .then(found => {
-                    let referencedFiles = found[0];
-                    let foundFiles = found[1];
-
-                    // for each file, check whether it is in the list of referenced files
-                    let promises: Promise<InlineString>[] = [];
-
-                    foundFiles.forEach((file, index, array) => {
-                        let foundFile: vscode.Uri | undefined = referencedFiles.find(match => match.fsPath === file.fsPath);
-                        if (J.Util.isNullOrUndefined(foundFile)) {
-                            this.ctrl.logger.debug("injectAttachementLinks() - File link not present in entry: ", file);
-                            // files.push(file); 
-                            // we don't execute yet, just collect the promises
-                            promises.push(this.buildReference(doc, file));
-
-                        }
-                    });
-                    return Promise.all(promises);
-                })
-                .then((inlineStrings: InlineString[]) => {
-                    this.ctrl.logger.trace("injectAttachementLinks() - Number of references to synchronize: ", inlineStrings.length);
-
-                    if (inlineStrings.length > 0) {
-                        this.injectInlineString(inlineStrings[0], ...inlineStrings.splice(1));
-                    }
-
-                    resolve(doc);
-
-                })
-                .catch((err: Error) => {
-                    this.ctrl.logger.error("Failed to synchronize page with notes folder.", err);
-                    reject(err);
-                });
-        });
-    }
+   
 
 }
