+++ src/ext/index.ts
@@ -23,8 +23,6 @@ export { Commands, JournalCommands } from './commands';
 export { Startup } from './startup';
 export {
     Configuration, 
-    InlineTemplate, 
-    ScopedTemplate, 
-    HeaderTemplate, 
+
     SCOPE_DEFAULT
 } from './conf';
