+++ src/extension.ts
@@ -37,7 +37,7 @@ export function activate(context: vscode.ExtensionContext) {
         .then((ctrl) => journalStartup.registerCommands(ctrl, context))
         .then((ctrl) => journalStartup.registerCodeLens(ctrl, context))
         .then((ctrl) => journalStartup.registerSyntaxHighlighting(ctrl))
-      
+                
         .then((ctrl) => { 
             journalConfiguration = ctrl.configuration;
