+++ src/features/index.ts
@@ -1,1 +1,3 @@
 export { InputMatcher } from "./InputMatcher";
+export { NoteLinksSync } from "./NoteLinksSync";
+export { NoteLoader } from "./NoteLoader";
