+++ src/model/config.ts
@@ -1,0 +1,29 @@
+export enum JournalPageType {
+    note,
+    entry,
+    attachement
+}
+
+export interface ScopedTemplate {
+    name?: string;
+    scope?: string;
+    template: string;
+    value?: string;
+}
+
+
+
+export interface FilePattern extends ScopedTemplate {
+    type: JournalPageType;
+}
+
+export interface PathTemplate extends ScopedTemplate {
+    type: JournalPageType;
+}
+
+export interface HeaderTemplate extends ScopedTemplate {
+}
+
+export interface InlineTemplate extends ScopedTemplate {
+    after: string;
+}
