+++ src/model/inline.ts
@@ -1,0 +1,8 @@
+import * as vscode from 'vscode';
+
+export interface InlineString {
+    position: vscode.Position;
+    value: string;
+    document: vscode.TextDocument;
+
+}
