+++ src/test/suite/input.test.ts
@@ -1,5 +1,5 @@
 import * as assert from 'assert';
-import moment = require('moment');
+
 
 // You can import and use all API from the 'vscode' module
 // as well as import your extension to test it
@@ -10,7 +10,7 @@ import { TestLogger } from '../TestLogger';
 suite('Open Journal Entries', () => {
 	vscode.window.showInformationMessage('Start all tests.');
 
-	/*
+	/* */
 	test('Simple', async () => {
 		assert.strictEqual(-1, [1, 2, 3].indexOf(5));
 		assert.strictEqual(-1, [1, 2, 3].indexOf(0));
@@ -138,56 +138,5 @@ suite('Open Journal Entries', () => {
 	});
 
 
-	test("Input 'w13'", async () => {
-		let config: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration("journal");
-		let ctrl = new J.Util.Ctrl(config);
-		ctrl.logger = new TestLogger(false);
 
-		let parser = new J.Actions.Parser(ctrl);
-		let input = await parser.parseInput("w13");
-
-		assert.ok(! input.hasOffset(), "Offset is set, is " + input.offset);
-		assert.ok(! input.hasFlags(), "Input has flags " + JSON.stringify(input));
-		assert.ok(! input.hasTask(), "Input has task flag " + JSON.stringify(input));
-		assert.ok(! input.hasText(), "Input has no text " + JSON.stringify(input));
-		assert.ok(input.hasWeek(), "Input has no week definition " + JSON.stringify(input));
-	});
-*/
-
-	test("Input 'w'", async () => {
-		let config: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration("journal");
-		let ctrl = new J.Util.Ctrl(config);
-		ctrl.logger = new TestLogger(false);
-
-		let parser = new J.Actions.Parser(ctrl);
-		let input = await parser.parseInput("w");
-		let currentWeek = moment().week();
-
-		assert.ok(!input.hasOffset(), "Offset is set, is " + input.offset);
-		assert.ok(!input.hasFlags(), "Input has flags " + JSON.stringify(input));
-		assert.ok(!input.hasTask(), "Input has task flag " + JSON.stringify(input));
-		assert.ok(!input.hasText(), "Input has no text " + JSON.stringify(input));
-		assert.ok(input.hasWeek(), "Input has no week definition " + JSON.stringify(input));
-
-		assert.strictEqual(input.week, currentWeek, "weeks mismatch");
-	});
-
-	test("Input 'next week'", async () => {
-		let config: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration("journal");
-		let ctrl = new J.Util.Ctrl(config);
-		ctrl.logger = new TestLogger(false);
-
-		let parser = new J.Actions.Parser(ctrl);
-		let input = await parser.parseInput("next week");
-
-		let currentWeek = moment().week();
-
-		assert.ok(!input.hasOffset(), "Offset is set, is " + input.offset);
-		assert.ok(!input.hasFlags(), "Input has flags " + JSON.stringify(input));
-		assert.ok(!input.hasTask(), "Input has task flag " + JSON.stringify(input));
-		assert.ok(!input.hasText(), "Input has no text " + JSON.stringify(input));
-		assert.ok(input.hasWeek(), "Input has no week definition " + JSON.stringify(input));
-
-		assert.strictEqual(input.week, currentWeek + 1, "weeks mismatch");
-	});
 });
