+++ test/Readme.md
@@ -1,0 +1,6 @@
+# Test workspaces
+
+ * ws_manual -> default workspace for extension tests
+ * ws_unittests -> workspace for the unit tests (has to be cleaned before each test)
+
+Both are configured in launch.json
