--- test/direct.d.ts
@@ -1,2 +1,0 @@
-import { ScopedTemplate } from "../src/ext/index";
-export declare function replaceDateFormats(st: ScopedTemplate, date: Date): void;
