--- test/workspace/.gitignore
@@ -1,3 +1,0 @@
-journal
-private
-website
