--- test/workspace/.vscode/settings.json
@@ -1,77 +1,0 @@
-{
-  "journal.base": "${workspaceFolder}/journal",
-  "journal.dev": true,
-  "journal.locale": "en-US",
-  "journal.openInNewEditorGroup": false,
-  "journal.patterns": {
-    "entries": {
-        "path": "${base}\\${year}\\@Drafts\\Journal",
-        "file": "${month}-${day}.${ext}"
-    },
-    "notes": {
-        "path": "${base}\\${year}\\@Drafts",
-        "file": "${input}.${ext}"
-    },
-},
-  "journal.templates": [
-    {
-      "name": "memo", 
-      "template": "- ${d:HHmm}: ${input}", 
-      "after": "## Notes"
-    }, 
-    {
-      "name": "task", 
-      "template": "- [] Task: ${input}", 
-      "after": "## Tasks"
-    }, 
-    {
-      "name": "entry", 
-      "template": "# ${d:dddd, MMMM DD YYYY}\n\n## Tasks\n\n## Notes\n\n"
-    }, 
-    {
-      "name": "time", 
-      "template": "${localTime}"
-    },     
-    {
-      "name": "files",
-      "template": "- NOTE: [${title}](${link})", 
-      "after": "## Notes"
-    }
-  ]
-  , 
-
-  "journal.scopes": [
-    {
-      "name": "me", 
-      "patterns": {
-        "notes": {
-          "path": "${base}\\${year}\\@Drafts\\Personal",
-          "file": "${d: YYYY-MM-DD}-${input}.${ext}"
-        }, 
-      }
-  }, 
-     {
-      "name": "private", 
-      "base": "${workspaceRoot}/private", 
-      "patterns": {
-        "notes": {
-          "path": "${base}",
-          "file": "${d:YYYYY-DD-MM}-${input}.${ext}"
-        }
-      }
-    }, 
-    {
-      "name": "website", 
-      "base": "${workspaceRoot}/website", 
-      "patterns": {
-        "notes": {
-          "path": "${base}",
-          "file": "${month}-${input}.${ext}"
-        }
-      }
-    }
-  ],
-  "files.exclude": {
-    ".*": true
-  }
-}
