+++ test/ws_manual/.gitignore
@@ -1,0 +1,3 @@
+*
+!.vscode
+!.gitignore
