+++ test/ws_unittests/.gitignore
@@ -1,0 +1,3 @@
+*
+!.vscode
+!.gitignore
